requirements.txt
# Web Scraping et Requests
beautifulsoup4
selenium
selenium-wire
fake-useragent
requests
webdriver-manager

# Data Analysis et Data Science
pandas
numpy
scipy
scikit-learn
statsmodels

# Visualisation
seaborn
matplotlib
plotly
plotly-express


# Interface utilisateur
streamlit
streamlit-extras

# Deep Learning et LLM (optionnel)
tensorflow
keras
autogen
dspy
# Traitement du texte
nltk
spacy

# Utilitaires
webdriver-manager
psutil
python-dotenv
tqdm

# Cloud et Base de données
google-cloud-bigquery
SQLAlchemy

# Utilitaires
PyYAML
typing-extensions
pathlib
psutil
